+++ Polaris/Tasks/Delete_Tasks.cs
@@ -1,7 +1,7 @@
 ﻿using System.Drawing;
 using System.Windows.Forms;
 
-namespace Polaris
+namespace Polaris.Tasks
 {
     public partial class Delete_Tasks : Form
     {
@@ -10,13 +10,13 @@ namespace Polaris
             InitializeComponent();
         }
 
-        private void panel1_Paint(object sender, PaintEventArgs e)
+        private void MenuPanel_Paint(object sender, PaintEventArgs e)
         {
-            ControlPaint.DrawBorder(e.Graphics, ViewHeader.ClientRectangle,
-            ColorTranslator.FromHtml("#27282F"), 0, ButtonBorderStyle.Solid,
-            ColorTranslator.FromHtml("#27282F"), 2, ButtonBorderStyle.Solid,
-            ColorTranslator.FromHtml("#27282F"), 0, ButtonBorderStyle.Solid,
-            ColorTranslator.FromHtml("#27282F"), 0, ButtonBorderStyle.Solid);
+            ControlPaint.DrawBorder(e.Graphics, MenuPanel.ClientRectangle,
+                ColorTranslator.FromHtml("#27282F"), 0, ButtonBorderStyle.Solid,
+                ColorTranslator.FromHtml("#27282F"), 2, ButtonBorderStyle.Solid,
+                ColorTranslator.FromHtml("#27282F"), 0, ButtonBorderStyle.Solid,
+                ColorTranslator.FromHtml("#27282F"), 0, ButtonBorderStyle.Solid);
         }
     }
 }
