+++ Polaris/Tasks/Draft_Tasks.cs
@@ -1,7 +1,7 @@
 ﻿using System.Drawing;
 using System.Windows.Forms;
 
-namespace Polaris
+namespace Polaris.Tasks
 {
     public partial class Drafts_Tasks : Form
     {
@@ -10,9 +10,9 @@ namespace Polaris
             InitializeComponent();
         }
 
-        private void ViewHeader_Paint(object sender, PaintEventArgs e)
+        private void MenuPanel_Paint(object sender, PaintEventArgs e)
         {
-            ControlPaint.DrawBorder(e.Graphics, ViewHeader.ClientRectangle,
+            ControlPaint.DrawBorder(e.Graphics, MenuPanel.ClientRectangle,
                 ColorTranslator.FromHtml("#27282F"), 0, ButtonBorderStyle.Solid,
                 ColorTranslator.FromHtml("#27282F"), 2, ButtonBorderStyle.Solid,
                 ColorTranslator.FromHtml("#27282F"), 0, ButtonBorderStyle.Solid,
